+++ clj/CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file. This change
 - `income-statement` query for profit and loss
 - diagnostics for failed balance directives show all the sub-accounts
 - f/none-f combinator filter, the opposite of f/some-f
+- validation of directives against Clojure spec
 
 ### Changed
 
@@ -20,6 +21,7 @@ All notable changes to this project will be documented in this file. This change
 - journal now includes posting flags, if any
 - only `*beans*` is used by the top-level queries, with `*directives*` et al being ignored
 - exception stack traces are no longer printed, instead they are saved in `*exception*`
+- testing support library broken out into `limabean-test` in `clj-test`, for external use
 
 [commit log]: https://github.com/tesujimath/limabean/compare/0.4.1...HEAD